refactoring : add comment and rename
